@@ -51,7 +51,7 @@
     "webpack-hot-middleware": "^2.9.1"
   },
   "dependencies": {
-    "react": "^0.14.7",
-    "react-dom": "^0.14.7"
+    "react": "^15.0.0-rc.2",
+    "react-dom": "^15.0.0-rc.2"
   }
 }
